feat(exec-core): extract pure helpers into wiredesk-exec-core crate (Task 1)
New workspace crate `wiredesk-exec-core` owns the sentinel-driven
exec helpers (format_command, parse_sentinel, clean_stdout,
parse_ready, format_timeout_diagnostic, strip_ansi,
is_powershell_prompt, is_remote_prompt) and shared types
(ShellKind, OneShotState). 32 helper unit tests migrated alongside
the code; wiredesk-term now imports them through the crate.

Foundation for Task 2-3 (ExecTransport trait + run_oneshot
extraction with streaming) and Task 6 (GUI IPC handler reusing the
same logic). uuid lifted to a workspace dependency so every
consuming crate sees the same pinned version.

Net code: -664 lines from term/main.rs (helpers + tests),
+645 lines split across the new crate. Workspace tests:
149 client + 97 host + 60 protocol + 19 term + 32 exec-core
+ 4 transport = 361 (was 361 before extraction; only locations
changed). cargo clippy --workspace -- -D warnings clean.
