+++ crates/wiredesk-exec-core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "wiredesk-exec-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+thiserror.workspace = true
+log.workspace = true
+uuid.workspace = true
